# src/specs/openapi.yaml
openapi: 3.1.0
info:
  title: geocontext
  description: |-
    Cette API **expérimentale** vise à faciliter l'exploitation des données de la Géoplateforme avec des LLM.
    Elle a vocation à devenir un module API Carto facilitant l'utilisation 
    des données de la Géoplateforme si son utilité est confirmée.

    NB : Il convient en parallèle d'intégrer l'appel à un service de géocodage 
    (voir [gpf-autocompletion.yaml](gpf-autocompletion.yaml) pour intégrer celui de la Géoplateforme)
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.5.9

servers:
 - url: http://localhost:3000

termsOfService: https://github.com/mborne/geocontext#readme
contact:
  email: apicarto@ign.fr
externalDocs:
  description: Dépôt GitHub
  url: https://github.com/mborne/geocontext#readme

tags:
  - name: geocode
    description: Géocodage
  - name: altitude
    description: Altitude
  - name: administratif
    description: Unités administratives
  - name: urbanisme
    description: Géoportail de l'Urbanisme

paths:

  #---------------------------------------------------
  # /v1/geocode
  #---------------------------------------------------
  /v1/geocode:
    get:
      summary: Renvoie les coordonnées (lon,lat) d'un lieu en complétant les informations.
      description: Le service service d'autocompletion de la Géoplateforme est appelé avec le texte en paramètres.
      operationId: geocode
      tags:
        - geocode
      parameters:
        - name: text
          in: query
          description: Le texte devant être completé
          required: true
          schema:
            type: string
          example: "2, avenue pasteur saint m"
      responses:
        "200":
          description: Une liste de suggestions de completions possibles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeocodeResponse"
              example:
                [
                  {
                    "country": "StreetAddress",
                    "city": "Saint-Malo",
                    "x": -2.004141,
                    "y": 48.655722,
                    "zipcode": "35400",
                    "street": "2 av pasteur",
                    "classification": 7,
                    "kind": "",
                    "fulltext": "2 av pasteur,35400 Saint-Malo",
                  },
                  {
                    "country": "PositionOfInterest",
                    "names": [
                      "Saint-Malo"
                    ],
                    "zipcode": "35400",
                    "zipcodes": [
                      "35400"
                    ],
                    "metropole": true,
                    "city": "Saint-Malo",
                    "street": "Saint-Malo",
                    "poiType": [
                      "lieu-dit habité",
                      "zone d'habitation"
                    ],
                    "kind": "lieu-dit habité",
                    "fulltext": "Saint-Malo, 35400 Saint-Malo",
                    "classification": 7,
                    "x": -1.994067,
                    "y": 48.647186
                  }
                ]
        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/altitude:
    get:
      tags:
        - altitude
      summary: Renvoie l’altitude pour une position donnée par sa longitude et sa latitude.
      description: Le service d'altimétrie de la Géoplateforme est appelé avec la longitude et la latitude en paramètres.
      operationId: getAltitude
      parameters:
        - in: query
          name: lon
          required: true
          description: La longitude du point.
          example: 6.869433
          schema:
            type: number
        - in: query
          name: lat
          required: true
          description: La latitude du point.
          example: 45.923697
          schema:
            type: number
      responses:
        '200':
          description: Position avec l'altitude.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AltitudeItem'
        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /v1/adminexpress:
    get:
      tags:
        - administratif
      summary: Renvoie les unités administratives (commune, département, région,...) pour une position donnée par sa longitude et sa latitude.
      description: |
        Le service WFS de la Géoplateforme (ADMINEXPRESS-COG.LATEST) est interrogé pour 
        récupérer les types suivants par intersection avec le point en entrée.
          * commune
          * canton
          * collectivite_territoriale
          * epci
          * departement
          * region 
          * arrondissement
      operationId: getAdminUnitsFromAdminExpress
      parameters:
        - in: query
          name: lon
          required: true
          description: the longitude of a location
          example: 6.869433
          schema:
            type: number
        - in: query
          name: lat
          required: true
          description: the latitude of a location
          example: 45.923697
          schema:
            type: number
      responses:
        '200':
          description: Unités administratives trouvées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminExpressItem'

        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /v1/parcellaire-express:
    get:
      tags:
        - administratif
      summary: "Renvoie les informations cadastrales pour une position donnée par sa longitude et sa latitude."
      description: |
        Le service WFS de la Géoplateforme (CADASTRALPARCELS.PARCELLAIRE_EXPRESS) est interrogé pour 
        récupérer les informations à proximité du point en entrée.
        
        Remarques :
        - Le résultat le plus proche est conservé pour chaque type
        - Une distance de 0.0 correspond au cas où il y a intersection.
      operationId: getParcellaireExpress
      parameters:
        - in: query
          name: lon
          required: true
          description: La longitude
          example: 6.869433
          schema:
            type: number
        - in: query
          name: lat
          required: true
          description: La latitude
          example: 45.923697
          schema:
            type: number
      responses:
        '200':
          description: Informations cadastrales trouvées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParcellaireExpressItem'

        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'



  /v1/urbanisme:
    get:
      tags:
        - urbanisme
      summary: "Renvoie les informations du document d'urbanisme (PLU, PLUi, POS, CC, PSMV) pour une position donnée par sa longitude et sa latitude."
      description: |
        Le service WFS de la Géoplateforme (wfs_du et wfs_scot) est interrogé pour 
        récupérer les informations suivantes à proximité du point en entrée.
        La distance au point de recherche est ajoutée dans la réponse 
        (0.0 correspond au cas où il y a intersection).
      operationId: getUrbanisme
      parameters:
        - in: query
          name: lon
          required: true
          description: La longitude
          example: 6.869433
          schema:
            type: number
        - in: query
          name: lat
          required: true
          description: La latitude
          example: 45.923697
          schema:
            type: number
      responses:
        '200':
          description: Documents d’urbanisme trouvés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UrbanismeItem'

        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /v1/sup:
    get:
      tags:
        - urbanisme
      summary: "Renvoie les assiettes des servitudes d'utilité publique (SUP) pour une position donnée par sa longitude et sa latitude."
      description: |
        Le service WFS de la Géoplateforme (wfs_sup) est interrogé pour 
        récupérer les types suivants à proximité du point en entrée.
        La distance au point de recherche est ajoutée dans la réponse (0.0 correspond au cas 
        où il y a intersection avec l'objet trouvé).
      operationId: getAssiettesServitudes
      parameters:
        - in: query
          name: lon
          required: true
          description: La longitude
          example: 6.869433
          schema:
            type: number
        - in: query
          name: lat
          required: true
          description: La latitude
          example: 45.923697
          schema:
            type: number
      responses:
        '200':
          description: Assiettes des servitudes d'utilité publique trouvées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssietteServitudeItem'

        '400':
          description: Mauvaise requête (paramètre invalide)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '500':
          description: Erreur système
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          examples: ["An error occurred"]
    BadRequestResponse:
      type: object
      properties:
        message:
          type: string
          examples: ["Invalid request parameters"]



    #---------------------------------------------------
    # /v1/altitude
    #---------------------------------------------------
    AltitudeItem:
      type: object
      properties:
        source:
          type: string
          examples: ["Géoplateforme (altimétrie)"]
        lon:
          type: number
          description: La longitude du point.
        lat:
          type: number
          description: La latitude du point.
        altitude:
          type: number
          description: L'altitude du point en mètres (null ou -99999 en l'absence de résultat).
        accuracy:
          type: string
          description: La précision de la valeur au point considéré.

    #---------------------------------------------------
    # /v1/adminexpress
    #---------------------------------------------------
    AdminExpressItem:
      type: object
      properties:
        type:
          type: string
          description: Le type de l'unité administrative.
          enum:
            - commune
            - canton
            - collectivite_territoriale
            - epci
            - departement
            - region
            - arrondissement
        id:
          type: string
          description: L'identifiant BDTOPO.
          examples: ["COMMUNE_0000000009750652"]
        bbox:
          type: array
          schema:
            $ref: '#/components/schemas/BoundingBox'
        nom_officiel:
          type: string
          description: Le nom de l'unité administrative.
          examples: ["Chamonix-Mont-Blanc"]
        nom_officiel_en_majuscules:
          type: string
          description: Le nom de l'unité administrative en majuscule.
          examples: ["CHAMONIX-MONT-BLANC"]
        code_insee:
          type: string
          description: Le code INSEE de l'unité administrative.
          examples: ["74056"]
        code_insee_du_departement:
          type: string
          description: Le code INSEE du département.
          examples: ["74"]
        code_insee_de_la_region:
          type: string
          description: Le code INSEE de la région.
          examples: ["74"]
        statut:
          type: string
          description: Le statut de la commune.
          examples: ["Commune simple"]
        population:
          type: number
          description: La population de la commune (non disponible pour les départements et régions)
          examples: [8642]
        date_du_recensement:
          type: string
          description: La date du recensement
          examples: ["2022-01-01Z"]
        organisme_recenseur:
          type: string
          description: L'organisme recenseur
          examples: ["INSEE"]
        code_insee_du_canton:
          type: string
          description: Le code INSEE du canton
          examples: ["10"]
        code_insee_de_l_arrondissement:
          type: string
          description: Le code INSEE de l'arrondissement
          examples: ["2"]
        codes_siren_des_epci:
          type: string
          description: Le code SIREN des EPCI
          examples: ["200023372"]

    #---------------------------------------------------
    # /v1/parcellaire-express
    #---------------------------------------------------
    ParcellaireExpressItem:
      type: object
      externalDocs:
        description: Parcellaire Express (PCI) - Descriptif de contenu
        url: https://geoservices.ign.fr/sites/default/files/2022-06/DC_Parcellaire_Express_%28PCI%29_1-1.pdf
      properties:
        id:
          type: string
          description: Identifiant technique
        type:
          type: string
          description: classe de l'objet
          enum:
            - arrondissement
            - commune
            - feuille
            - subdivision_fiscale
            - parcelle
            - localisant
        bbox:
          type: array
          schema:
            $ref: '#/components/schemas/BoundingBox'
        distance:
          type: number
          description: La distance en degrée entre l'objet et la position.
          minimum: 0
        source:
          type: string
          examples: ["Géoplateforme (WFS, CADASTRALPARCELS.PARCELLAIRE_EXPRESS)"]
        code_dep:
          type: string
          description: Code INSEE du département.
          examples: ["74"]
        code_com:
          type: string
          description: Code de la commune.
          examples: ["056"]
        nom_com:
          type: string
          description: Nom de la commune.
          examples: ["Chamonix-Mont-Blanc"]
        code_insee:
          type: string
          description: Numéro INSEE de la commune.
          examples: ["74056"]
        feuille:
          type: integer
          description: Numéro de la feuille cadastrale.
          examples: [7]
        section:
          type: string
          description: Numéro de la section cadastrale.
        com_abs:
          type: string
          description: Ancien code de la commune en cas de fusion de communes.
          examples: ["000"]
        code_arr:
          type: string
          description: Code de l’arrondissement.
          examples: ["000"]
        idu:
          type: string
          description: Identifiant de la parcelle cadastrale.
          examples: ["740560000G4353"]
        numero:
          type: string
          description: Numéro de la parcelle cadastrale.
          examples: ["4353"]
        contenance:
          type: integer
          minimum: 0
          description: Superficie extraite de MAJIC exprimée en m².
          examples: [178]
        echelle:
          type: string
          description: Dénominateur de l’échelle principale du plan cadastral.
          examples: [500]
        edition:
          type: string
          format: date
          description: Numéro d’édition de la feuille cadastrale
          examples: ["2023-07-01"]
      required:
        - id
        - type
        - bbox
        - nom_com
        - code_dep
        - code_insee
        - source

    #---------------------------------------------------
    # /v1/urbanisme
    #---------------------------------------------------
    UrbanismeItem:
      type: object
      properties:
        type:
          type: string
          description: Identifiant du type de l'objet
          enum:
            - scot
            - document
            - zone_urba
            - secteur_cc
            - info_pct
            - info_lin
            - info_surf
            - prescription_pct
            - prescription_lin
            - prescription_surf
        id:
          type: string
          description: Identifiant technique de l'objet
        bbox:
          type: array
          schema:
            $ref: '#/components/schemas/BoundingBox'
        distance:
          type: number
          description: La distance en mètre entre l'objet et la position.
          minimum: 0
        source:
          type: string
          examples: ["Géoplateforme - (WFS Géoportail de l'Urbanisme)"]
        gid:
          type: number
          description: Identifiant technique de l'objet
          examples: [2905619]
        gpu_doc_id:
          type: string
          description: Identifiant unique du document sur le GpU
          examples: ["398a91a8349876bc9f08c2830347b4f1"]
        partition:
          type: string
          description: Identifiant stable du document d'urbanisme sur le GpU
          examples: ["DU_74056"]
        # champs métiers
        libelle:
          type: string
          description: Description physique de la prescription ou du périmètre d’information.
          examples: ["Zone à risque d'exposition au plomb"]
        libelong:
          type: string
          description: Nom complet littéral de la zone tel qu’il apparaît dans le chapitre du règlement écrit.
        typezone:
          type: string
          description: Type de la zone classé dans une nomenclature simplifiée.
        txt:
          type: string
          description: Étiquette associée à la prescription ou à l'information
        typeinf:
          type: string
          description: Type d'information (InformationUrbaType)
          examples: ["17"]
        stypeinf:
          type: string
          description: sous-type détaillant le type d'information
          examples: ["00"]
        idurba: 
          type: string
          description: Identifiant du document d'urbanisme
          examples: ["74056_PLU_20231010"]
        datvalid:
          type: string
          description: Date de la dernière validation de la prescription.
        nomfic:
          type: string
          description: Nom du fichier pdf contenant le texte décrivant la prescription
          examples: ["74056_reglement_20231010.pdf"]
        # TODO : compléter ( https://cnig.gouv.fr/IMG/pdf/231220_standard_cnig_plu_v2024-01.pdf )
        # "typep": null,
        # "urlfic": null,
        # "insee": null,
        # "datvalid": null,
        # "stypep": null,
        # "idinfo": null,
        # "lib_idinfo": null,
        # "nature": null,
        # "symbole": null
      required:
        - type
        - id
        - bbox
        - source



    #---------------------------------------------------
    # /v1/assiette-sup
    #---------------------------------------------------
    AssietteServitudeItem:
      type: object
      description: Assiette de servitude d'utilité publique
      properties:
        type:
          type: string
          description: Identifiant du type de l'objet
          enum:
            - assiette_sup_p
            - assiette_sup_l
            - assiette_sup_s
        id:
          type: string
          description: Identifiant technique de l'objet
        bbox:
          type: array
          schema:
            $ref: '#/components/schemas/BoundingBox'
        distance:
          type: number
          description: La distance en mètre entre l'objet et la position.
          minimum: 0
        source:
          type: string
          examples: ["Géoplateforme - (WFS Géoportail de l'Urbanisme)"]
        gid:
          type: number
          description: Identifiant technique de l'objet
          examples: ["2905619"]
        gpu_doc_id:
          type: string
          description: Identifiant unique du document sur le GpU
          examples: ["398a91a8349876bc9f08c2830347b4f1"]
        suptype:
          type: string
          description: Code du type de servitude
          examples: ["ac1"]
        partition:
          type: string
          description: Identifiant stable du document sur le GpU
          examples: ["172014607_SUP_74_AC1"]
        # champs métiers
        fichier:
          type: string
          description: Nom ou référence du ficher contenant l'acte instituant la servitude
          examples: ["AC1_Deux-fontaines--Eglise-Saint-Michel--Presbytere_20230421_act.pdf"]
        idass:
          type: string
          description: Identifiant de l'assiette
          examples: ["AC1-172014607-2304250003-1-1"]
        idgen:
          type: string
          description: Identifiant du générateur
          examples: ["AC1-172014607-2304250003-1"]
        nomass:
          type: string
          description: Nom abrégé de l'assiette
        typeAss:
          type: string
          description: Nature de l'assiette selon sa vocation principale et la catégorie de SUP
          examples: ["Périmètre des abords"]
        modegeoass:
          type: string
          description: Description de la méthode utilisée pour générer la géométrie de l'assiette
          examples: ["Égale au générateur"]
        paramcalc:
          type: number
          description: Valeur du paramètre ayant permis de calculer l'assiette lorsque celle-ci correspond à un objet tampon
        srcgeoass:
          type: number
          description: Type de carte, de référentiel géographique utilisé en tant que source de géoréférencement lors de la digitalisation de la géométrie
          examples: ["Cadastre"]
        datesrcass:
          type: string
          description: Date d'actualité du référentiel utilisé lors du géoréférencement de l'objet
          examples: ["20180101"]
        # "angle1": null,
        # "angle2": null,
        # "rayon": null,
        # "h": null,
        # "href": null,
        # "xdebut": null,
        # "xfinal": null,
        # "ydebut": null,
        # "yfinal": null,
        # "largeur": null,
        nomsuplitt:
          type: string
          description: Nom littéral de la servitude, figurant dans l'acte l'ayant instaurée
          examples: ["Deux fontaines ; Eglise Saint-Michel ; Presbytère"]
      required:
        - type
        - id
        - bbox
        - source

    #--------------------------------------------------------------------------
    # geojson
    #--------------------------------------------------------------------------
    BoundingBox:
      description: "Geo-spatial bounding box."
      type: array
      items:
        type: number
      minItems: 4
      maxItems: 4


    #---------------------------------------------------
    # /v1/geocode
    #---------------------------------------------------

    GeocodeResponse:
      type: array
      items:
        oneOf:
          - $ref: "#/components/schemas/GeocodeAddress"
          - $ref: "#/components/schemas/GeocodePoi"

    GeocodeAddress:
      type: object
      required:
        - country
        - city
        - x
        - y
        - zipcode
        - street
        - classification
        - kind
        - fulltext
        - metropole
      properties:
        country:
          type: string
        city:
          type: string
        x:
          type: number
          format: float
        y:
          type: number
          format: float
        zipcode:
          type: string
          minLength: 5
          maxLength: 5
        street:
          type: string
        classification:
          type: integer
          format: int32
        kind:
          type: string
        fulltext:
          type: string
        metropole:
          type: boolean
    GeocodePoi:
      type: object
      required:
        - country
        - city
        - x
        - y
        - zipcode
        - zipcodes
        - poiType
        - street
        - classification
        - kind
        - fulltext
        - metropole
      properties:
        country:
          type: string
        city:
          type: string
        x:
          type: number
          format: float
        y:
          type: number
          format: float
        zipcode:
          type: string
          minLength: 5
          maxLength: 5
        zipcodes:
          type: array
          items:
            type: string
            minLength: 5
            maxLength: 5
        poiType:
          type: array
          items:
            type: string
        street:
          type: string
        classification:
          type: integer
          format: int32
        kind:
          type: string
        fulltext:
          type: string
        metropole:
          type: boolean
    GeocodeError:
      type: object
      required:
        - status
        - results
      properties:
        status:
          type: string
        results:
          type: array
          items:
            type: object